Move field and label blueprint to structure module
