+++ src/yafowil/tests/__init__.py
@@ -5,8 +5,10 @@ from yafowil import common
 from yafowil import compound
 from yafowil import datatypes
 from yafowil import persistence
+from yafowil import structure
 from yafowil import table
 from yafowil.base import factory
+from yafowil.utils import Tag
 import lxml.etree as etree
 import sys
 import unittest
@@ -22,6 +24,7 @@ class YafowilTestCase(NodeTestCase):
         reload(compound)
         reload(datatypes)
         reload(persistence)
+        reload(structure)
         reload(table)
 
     def tearDown(self):
@@ -34,6 +37,13 @@ def fxml(xml):
     return etree.tostring(et, pretty_print=True).decode('utf-8')
 
 
+def wrapped_fxml(value):
+    return fxml(u'<div>' + value + u'</div>')
+
+
+tag = Tag(lambda msg: msg)
+
+
 def test_suite():
     from yafowil.tests import test_base
     from yafowil.tests import test_common
@@ -42,6 +52,7 @@ def test_suite():
     from yafowil.tests import test_datatypes
     from yafowil.tests import test_persistence
     from yafowil.tests import test_resources
+    from yafowil.tests import test_structure
     from yafowil.tests import test_table
     from yafowil.tests import test_tsf
     from yafowil.tests import test_utils
@@ -55,6 +66,7 @@ def test_suite():
     suite.addTest(unittest.findTestCases(test_datatypes))
     suite.addTest(unittest.findTestCases(test_persistence))
     suite.addTest(unittest.findTestCases(test_resources))
+    suite.addTest(unittest.findTestCases(test_structure))
     suite.addTest(unittest.findTestCases(test_table))
     suite.addTest(unittest.findTestCases(test_tsf))
     suite.addTest(unittest.findTestCases(test_utils))
